Large scale refactoring.
I've made many changes which have resulted in a performance gain but at the
expense of accuracy.

Introduced a nubmer of tunable nobs to do things like downsize each frame before
processing, only count so-many colors, and what image filtering algorithm to
use.

Made the code style more consistent, as well as removing the args parsing (for
the time being, can be re-added later).

I still feel that more can be done, but for now this is what I've got.